+++ libraries/MCP4725/MCP4725.h
@@ -3,7 +3,7 @@
 //    FILE: MCP4725.h
 //  AUTHOR: Rob Tillaart
 // PURPOSE: Arduino library for 12 bit I2C DAC - MCP4725
-// VERSION: 0.3.7
+// VERSION: 0.3.8
 //     URL: https://github.com/RobTillaart/MCP4725
 //
 
@@ -12,7 +12,7 @@
 #include "Arduino.h"
 
 
-#define MCP4725_VERSION         (F("0.3.7"))
+#define MCP4725_VERSION         (F("0.3.8"))
 
 
 //  CONSTANTS
@@ -68,6 +68,14 @@ public:
   //  due to rounding the returned value can differ slightly.
   float    getPercentage();
 
+  //  typical and default value = 5.0 V
+  void     setMaxVoltage(float v = 5.0);
+  float    getMaxVoltage();
+  //  will set 0..5V to the closest integer value in range 0..4095
+  int      setVoltage(float v);
+  // returns (approximation of) the set voltage. Assumes linear mapping.
+  float    getVoltage();
+
 
   //  unfortunately it is not possible to write a different value
   //  to the DAC and EEPROM simultaneously or write EEPROM only.
@@ -99,6 +107,8 @@ private:
   uint8_t  _readRegister(uint8_t* buffer, const uint8_t length);
 
   int      _generalCall(const uint8_t gc);
+
+  float    _maxVoltage;
 
   TwoWire*  _wire;
 };